feat(monolith): add register_public hooks and knowledge public_router
Factor the two public knowledge handlers (/api/knowledge/public/graph and
/public/notes/{note_id}) into knowledge/public_router.py, which imports only the
neutral helpers (http_cache, service.get_vault_root) so the public import graph
never pulls in the private-route module. router.py drops the moved handlers and
now-unused helper defs and imports the shared helpers instead.

Add register_public(app) to each public-tier domain: ships/stars/hikes/dr_jobs
reuse register (wholly public); knowledge mounts only public_router; home mounts
only observability_router (schedule stays private). knowledge.register still
includes public_router so the private app serves /api/knowledge/public/* exactly
as before. Exclude the new register_public functions from the BDD coverage
contract (app wiring, mirrors the existing register exclusions).
