+++ projects/monolith/knowledge/router.py
@@ -14,17 +14,13 @@ can override it with a deterministic fake via ``app.dependency_overrides``.
 from __future__ import annotations
 
 import logging
-import os
 import re
-from datetime import datetime, timezone
 from email.utils import format_datetime
-from pathlib import Path
 from typing import Literal
 
 import yaml
 from fastapi import APIRouter, Depends, HTTPException, Query, Request, Response
 from pydantic import BaseModel
-from sqlalchemy import func, or_
 from sqlmodel import Session, select
 
 from app.db import get_session
@@ -40,10 +36,11 @@ from knowledge.gaps import (
     undelete_gap,
     verify_gap,
 )
-from knowledge.gardener import Gardener, _slugify
+from knowledge.gardener import Gardener
+from knowledge.http_cache import _as_utc, _graph_etag, _GRAPH_CACHE_CONTROL
 from knowledge.indexing import index_note_best_effort
 from knowledge.ingest_queue import IngestQueueItem, ingest_raw
-from knowledge.models import AtomRawProvenance, Note, NoteLink, RawInput
+from knowledge.models import AtomRawProvenance, RawInput
 from knowledge.notes import (
     _note_to_review_dict,
     delete_note,
@@ -54,14 +51,8 @@ from knowledge.notes import (
     undelete_note,
     verify_note_visibility,
 )
-from knowledge.service import DEFAULT_VAULT_ROOT, VAULT_ROOT_ENV
-from knowledge.store import GRAPH_NOTE_TYPES, KnowledgeStore
-from knowledge.visibility import (
-    effective_visibility,
-    public_notes_filter,
-    sanitize_public_body,
-)
-from knowledge.visibility import _slugify as _visibility_slugify
+from knowledge.service import get_vault_root
+from knowledge.store import KnowledgeStore
 from shared.embedding import EmbeddingClient
 
 logger = logging.getLogger(__name__)
@@ -76,11 +67,6 @@ def get_embedding_client() -> EmbeddingClient:
     to inject a deterministic fake.
     """
     return EmbeddingClient()
-
-
-def _get_vault_root() -> Path:
-    """Resolve the vault root from the env (or default), as an absolute path."""
-    return Path(os.environ.get(VAULT_ROOT_ENV, DEFAULT_VAULT_ROOT)).resolve()
 
 
 @router.get("/search")
@@ -111,37 +97,6 @@ async def search_knowledge(
     return {"results": results}
 
 
-# Mirrors NOTES_PAGE_CACHE_CONTROL in projects/monolith/frontend/src/lib/cache-headers.js — keep in sync.
-_GRAPH_CACHE_CONTROL = (
-    "public, s-maxage=3600, stale-while-revalidate=86400, stale-if-error=31536000"
-)
-
-
-def _as_utc(value: datetime | None) -> datetime | None:
-    """Coerce a datetime to tz-aware UTC.
-
-    Postgres returns tz-aware values; SQLite (used in tests) can return
-    naive ones even though we always write tz-aware UTC. Treat naive
-    datetimes as UTC so downstream formatters and ETag stamps are stable
-    across both backends.
-    """
-    if value is None:
-        return None
-    if value.tzinfo is None:
-        return value.replace(tzinfo=timezone.utc)
-    return value.astimezone(timezone.utc)
-
-
-def _graph_etag(node_count: int, indexed_at: datetime | None) -> str:
-    """Stable ETag for a graph payload.
-
-    Combines max(indexed_at) with node count so deletions invalidate even
-    when the surviving notes' timestamps don't move.
-    """
-    stamp = indexed_at.isoformat() if indexed_at is not None else "null"
-    return f'"{stamp}-{node_count}"'
-
-
 @router.get("/graph")
 def get_graph(
     request: Request,
@@ -169,184 +124,6 @@ def get_graph(
     return graph
 
 
-@router.get("/public/graph")
-def get_public_graph(
-    request: Request,
-    response: Response,
-    session: Session = Depends(get_session),
-):
-    """Public-only knowledge graph: only public nodes, only doubly-public edges.
-
-    Mirrors :func:`get_graph` but applies ``public_notes_filter()`` to both
-    ends of every edge so a private note can never appear as a node *or* a
-    target. Same Cache-Control + ETag semantics so the CDN treats this
-    payload identically.
-    """
-    # Only public notes whose type is in the renderable graph set. Gap
-    # stubs (type='gap') with NULL visibility are excluded by the
-    # visibility filter alone, but this also keeps the type filter
-    # consistent with get_graph().
-    public_note_rows = session.execute(
-        select(
-            Note.note_id,
-            Note.title,
-            Note.type,
-            Note.indexed_at,
-            # Prefer public-only layout positions (computed over the public
-            # subgraph by knowledge.service._run_public_layout_pass); fall
-            # back to the full-graph positions if the public pass hasn't
-            # populated this row yet, so a fresh deploy never serves NULL
-            # coords. Same shape consumed by /private/notes — keep both as
-            # nullable in the response, the client handles either.
-            func.coalesce(Note.layout_x_public, Note.layout_x).label("x"),
-            func.coalesce(Note.layout_y_public, Note.layout_y).label("y"),
-        )
-        .where(public_notes_filter())
-        .where(Note.type.in_(list(GRAPH_NOTE_TYPES)))
-        .where(Note.deleted_at.is_(None))
-    ).all()
-
-    public_note_ids = {row.note_id for row in public_note_rows}
-    slug_to_note_id = {_slugify(nid): nid for nid in public_note_ids}
-
-    if public_note_ids:
-        # Single SQL join enforces both-ends-public: source side via the
-        # join filter, target side via the IN clause against the resolved
-        # public slug set (mirrors get_graph's slug→canonical resolution).
-        link_rows = session.execute(
-            select(
-                Note.note_id.label("source"),
-                NoteLink.target_id.label("target"),
-                NoteLink.kind,
-                NoteLink.edge_type,
-            )
-            .join(Note, NoteLink.src_note_fk == Note.id)
-            .where(public_notes_filter())
-            .where(Note.deleted_at.is_(None))
-        ).all()
-    else:
-        link_rows = []
-
-    edges: list[dict] = []
-    for row in link_rows:
-        canonical_target = slug_to_note_id.get(_slugify(row.target))
-        if canonical_target is None:
-            continue
-        edges.append(
-            {
-                "source": row.source,
-                "target": canonical_target,
-                "kind": row.kind,
-                "edge_type": row.edge_type,
-            }
-        )
-
-    degree_by_note_id: dict[str, int] = {}
-    for edge in edges:
-        degree_by_note_id[edge["source"]] = degree_by_note_id.get(edge["source"], 0) + 1
-        degree_by_note_id[edge["target"]] = degree_by_note_id.get(edge["target"], 0) + 1
-
-    nodes = [
-        {
-            "id": row.note_id,
-            "title": row.title,
-            "type": row.type,
-            "degree": degree_by_note_id.get(row.note_id, 0),
-            "x": row.x,
-            "y": row.y,
-        }
-        for row in public_note_rows
-    ]
-
-    indexed_at = _as_utc(
-        max((row.indexed_at for row in public_note_rows), default=None)
-    )
-    etag = _graph_etag(len(public_note_rows), indexed_at)
-    headers = {"Cache-Control": _GRAPH_CACHE_CONTROL, "ETag": etag}
-    if indexed_at is not None:
-        headers["Last-Modified"] = format_datetime(indexed_at, usegmt=True)
-
-    if request.headers.get("if-none-match") == etag:
-        return Response(status_code=304, headers=headers)
-
-    for key, value in headers.items():
-        response.headers[key] = value
-    logger.info("public.graph.served nodes=%d edges=%d", len(nodes), len(edges))
-    return {
-        "nodes": nodes,
-        "edges": edges,
-        # Mirrors get_graph's response — StatusBar in the SvelteKit page
-        # reads this to display "indexed Xm ago". Previously omitted so
-        # the public page showed "indexed —" while the private one didn't.
-        "indexed_at": indexed_at.isoformat() if indexed_at is not None else None,
-    }
-
-
-@router.get("/public/notes/{note_id}")
-def get_public_note(
-    note_id: str,
-    session: Session = Depends(get_session),
-) -> dict:
-    """Return a single note iff its effective visibility is ``public``.
-
-    The 404 response is identical for missing notes AND for notes that
-    exist but are private/null-visibility — the existence of a private
-    note must never be observable. Body wikilinks targeting private
-    notes are stripped to plain text via :func:`sanitize_public_body`;
-    wikilinks targeting public notes are left intact for the frontend
-    renderer to resolve.
-    """
-    note = session.exec(
-        select(Note).where(Note.note_id == note_id).where(Note.deleted_at.is_(None))
-    ).one_or_none()
-    if note is None or effective_visibility(note) != "public":
-        # Identical 404 for missing and private — never expose existence.
-        # The reason is logged but not surfaced in the response.
-        reason = "not_found" if note is None else "private_gated"
-        logger.info("public.note.404 note_id=%s reason=%s", note_id, reason)
-        raise HTTPException(status_code=404, detail="Not Found")
-
-    # ADR 006 Phase 2: body of record is Postgres ``content``. The vault
-    # read is a fallback only while the Phase 1 backfill is in flight, and
-    # ``resolve_note_body`` keeps the same path-traversal guard so a
-    # malformed path can't escape the vault root.
-    vault_root = _get_vault_root()
-    body = resolve_note_body(note.content, note.path, vault_root)
-    if body is None:
-        # Same identical 404 — don't leak that the DB row exists but
-        # the body is unavailable.
-        logger.info("public.note.404 note_id=%s reason=vault_file_missing", note_id)
-        raise HTTPException(status_code=404, detail="Not Found")
-
-    # Slugified ids of every non-public note. ``sanitize_public_body``
-    # slugifies wikilink display text via ``visibility._slugify`` and
-    # drops the bracketing on any match — use the same helper here so
-    # the sets compare consistently. Note: ``or_(... != 'public', ...
-    # IS NULL)`` is required because in SQL ``NULL != 'public'`` is
-    # ``NULL``, not true, so a bare ``!=`` would silently leave NULL-
-    # visibility notes out of the private set.
-    private_slugs = {
-        _visibility_slugify(n.note_id)
-        for n in session.exec(
-            select(Note)
-            .where(or_(Note.visibility != "public", Note.visibility.is_(None)))
-            .where(Note.deleted_at.is_(None))
-        ).all()
-    }
-    sanitized = sanitize_public_body(body, private_slugs)
-
-    indexed_at = _as_utc(note.indexed_at)
-    logger.info("public.note.served note_id=%s", note_id)
-    return {
-        "note_id": note.note_id,
-        "title": note.title,
-        "tags": list(note.tags or []),
-        "aliases": list(note.aliases or []),
-        "indexed_at": indexed_at.isoformat() if indexed_at is not None else None,
-        "body": sanitized,
-    }
-
-
 @router.get("/notes/review-queue")
 def get_notes_review_queue_endpoint(
     mode: Literal["pending", "audit"] = Query(default="pending"),
@@ -367,7 +144,7 @@ def get_notes_review_queue_endpoint(
     ordering, ``GET /notes/review-queue`` would resolve to
     ``get_knowledge_note(note_id="review-queue")`` and 404.
     """
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     return {
         "notes": list_notes_for_review(
             session, mode=mode, limit=limit, vault_root=vault_root
@@ -387,7 +164,7 @@ def get_knowledge_note(
 
     # ADR 006 Phase 2: body of record is Postgres ``content``; the vault
     # read is a fallback only while the Phase 1 backfill is in flight.
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     body = resolve_note_body(note.get("content"), note["path"], vault_root)
     if body is None:
         raise HTTPException(status_code=404, detail="vault file missing")
@@ -414,7 +191,7 @@ def delete_note_endpoint(
     is now the standard review-dict payload (matching the other
     note-action endpoints), with ``deleted_at`` populated.
     """
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         note = delete_note(session, note_id, vault_root)
     except ValueError as exc:
@@ -435,7 +212,7 @@ def undelete_note_endpoint(
     Use the audit-mode review queue + delete action to land here in the
     first place.
     """
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         note = undelete_note(session, note_id, vault_root)
     except ValueError as exc:
@@ -456,7 +233,7 @@ async def edit_note(
     if note is None:
         raise HTTPException(status_code=404, detail="note not found")
 
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     resolved = (vault_root / note["path"]).resolve()
     if not resolved.is_relative_to(vault_root) or not resolved.is_file():
         raise HTTPException(status_code=404, detail="vault file missing")
@@ -680,7 +457,7 @@ def get_review_queue_endpoint(
     plus a ``stub_body`` read from ``_researching/<slug>.md`` so the
     audit UI can render in-place without a per-row round-trip.
     """
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     return {
         "gaps": list_gaps_for_review(
             session, mode=mode, limit=limit, vault_root=vault_root
@@ -721,7 +498,7 @@ def answer_gap_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Commit a user answer for a gap and emit a personal-tier atom."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         return answer_gap(session, gap_id, data.answer, vault_root)
     except ValueError as exc:
@@ -738,7 +515,7 @@ def reject_gap_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Reject a pending gap. Transitions in_review → rejected and tombstones the stub."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         return reject_gap(session, gap_id, vault_root)
     except ValueError as exc:
@@ -757,7 +534,7 @@ def approve_gap_endpoint(
     to ``classified`` so the next reconciler tick doesn't revert the
     Gap row. Only valid for ``gap_class='external'``.
     """
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         return approve_gap(session, gap_id, vault_root)
     except ValueError as exc:
@@ -794,7 +571,7 @@ def delete_gap_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Soft-delete a gap. Hard-deletes the stub file; regenerable on undelete."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         return delete_gap(session, gap_id, vault_root)
     except ValueError as exc:
@@ -847,7 +624,7 @@ def set_note_visibility_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Set ``visibility`` (public|private) and mark verified."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         note = set_note_visibility(session, note_id, data.visibility, vault_root)
     except ValueError as exc:
@@ -861,7 +638,7 @@ def verify_note_visibility_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Mark ``visibility_verified=True``. 409 if visibility is unset."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         note = verify_note_visibility(session, note_id)
     except ValueError as exc:
@@ -875,7 +652,7 @@ def reset_note_visibility_endpoint(
     session: Session = Depends(get_session),
 ) -> dict:
     """Clear ``visibility`` and ``visibility_verified``. Sends back to pending."""
-    vault_root = _get_vault_root()
+    vault_root = get_vault_root()
     try:
         note = reset_note_visibility(session, note_id, vault_root)
     except ValueError as exc:
