+++ projects/monolith/stars/__init__.py
@@ -18,6 +18,11 @@ def register(app: FastAPI) -> None:
     app.include_router(router)
 
 
+def register_public(app: FastAPI) -> None:
+    """Stars is a wholly public, read-only domain: reuse register."""
+    register(app)
+
+
 def on_startup_jobs(session: Session) -> None:
     from scheduler.api import register_job
     from stars.grid import load_climatology_handler, load_grid_handler
